updated search query builder updated attribute chooser to use trait objects instead of string labels/trait names

+++ app/src/main/java/com/fieldbook/tracker/database/dao/ObservationDao.kt
@@ -352,7 +352,7 @@ class ObservationDao {
             rep: String? = (getRep(studyId, plotId, traitDbId) + 1).toString()
         ): Long = withDatabase { db ->
 
-            val traitObj = ObservationVariableDao.getTraitById(traitDbId.toInt())
+            val traitObj = ObservationVariableDao.getTraitById(traitDbId)
             val internalTraitId = if (traitObj == null) {
                 -1
             } else {
